feat(byok): restrict available models to Anthropic and OpenAI for BYOK plans
Add BYOK_AVAILABLE_PROVIDER_IDS constant and filter out non-BYOK providers
in isModelAvailable. Lift filteredModels/modelsDescriptionByProvider
computation up to ModelProvidersPageContent so the filtering is applied
before passing provider descriptions to ProvidersToggleList.
