+++ front/components/pages/workspace/model_providers/ProvidersToggleList.tsx
@@ -1,20 +1,8 @@
 import { ProviderContextItem } from "@app/components/pages/workspace/model_providers/ProviderContextItem";
-import {
-  REASONING_MODEL_CONFIGS,
-  USED_MODEL_CONFIGS,
-} from "@app/components/providers/types";
-import { isModelCustomAvailable } from "@app/lib/assistant";
-import { useFeatureFlags } from "@app/lib/auth/AuthContext";
-import type {
-  ModelConfigurationType,
-  ModelProviderIdType,
-} from "@app/types/assistant/models/types";
+import type { ModelProviderIdType } from "@app/types/assistant/models/types";
 import type { PlanType } from "@app/types/plan";
 import type { ProvidersSelection } from "@app/types/provider_selection";
 import { ContextItem } from "@dust-tt/sparkle";
-import groupBy from "lodash/groupBy";
-import mapValues from "lodash/mapValues";
-import uniqBy from "lodash/uniqBy";
 import { type Dispatch, type SetStateAction, useCallback } from "react";
 
 interface ProviderContextItemProps {
@@ -22,6 +10,7 @@ interface ProviderContextItemProps {
   setProvidersSelection: Dispatch<SetStateAction<ProvidersSelection>>;
   isWorkspaceValidating: boolean;
   plan: PlanType;
+  modelsDescriptionByProvider: Partial<Record<ModelProviderIdType, string>>;
 }
 
 export function ProvidersToggleList({
@@ -29,26 +18,8 @@ export function ProvidersToggleList({
   setProvidersSelection,
   isWorkspaceValidating,
   plan,
+  modelsDescriptionByProvider,
 }: ProviderContextItemProps) {
-  const { featureFlags } = useFeatureFlags();
-
-  // Filter models based on feature flags and build modelProviders dynamically
-  const filteredModels = uniqBy(
-    [...USED_MODEL_CONFIGS, ...REASONING_MODEL_CONFIGS],
-    (m) => m.modelId
-  ).filter(
-    (model) =>
-      !model.isLegacy && isModelCustomAvailable(model, featureFlags, plan)
-  );
-
-  const modelsDescriptionByProvider: Partial<
-    Record<ModelProviderIdType, string>
-  > = mapValues(
-    groupBy(filteredModels, "providerId"),
-    (modelConfigurations: ModelConfigurationType[]) =>
-      modelConfigurations.map(({ displayName }) => displayName).join(", ")
-  );
-
   const toggleProvider = useCallback(
     (provider: ModelProviderIdType) => {
       setProvidersSelection((previousSelection: ProvidersSelection) => ({
